pyproject.toml
[project]
name = "rag-research-graph"
version = "0.0.1"
description = "Starter template for making a custom RAG research agent graph in LangGraph."
authors = [
    { name = "Henry Vu", email = "phuocvu2001@gmail.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.6",
    "langchain-openai>=0.2.14",
    "langchain-anthropic>=0.1.23",
    "langchain>=0.3.13",
    "langchain-fireworks>=0.1.7",
    "langchain-elasticsearch>=0.2.2",
    "langchain-pinecone>=0.1.3",
    "langchain-redis>=0.1.1",
    "langchain-mongodb>=0.1.9",
    "langchain-cohere>=0.2.4",
    "langgraph-checkpoint>=2.0.9",
    "trustcall>=0.0.26",
    "python-dotenv>=1.0.1",

]

[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["retrieval_graph", "index_graph", "shared", "memory"]
[tool.setuptools.package-dir]
"langgraph.templates.retrieval_graph" = "src/retrieval_graph"
"langgraph.templates.index_graph" = "src/index_graph"
"langgraph.templates.memory_graph" = "src/memory"
"retrieval_graph" = "src/retrieval_graph"
"index_graph" = "src/index_graph"
"memory" = "src/memory"
"shared" = "src/shared"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
[tool.pytest.ini_options]
pythonpath = [
  "src"
]